requirements.txt
Telethon==1.28.5
Pillow==9.5.0
cryptg==0.4.0
aiohttp==3.8.4
python-dotenv==1.0.0